+++ src/function.c
@@ -21,27 +21,27 @@ int LoadGameFlag( char *fn );
 int SaveGameFlag( char *fn );
 int LoadGameFlag2( char *fn );
 int SaveGameFlag2( char *fn );
-int SaveFile( char *fn, long *buff, long size );
-int LoadFile( char *fn, long *buff, long size );
-long GetConfig( char* fn, char* cParam );
-long LogFileWrite( char* fn, char* cParam );
+int SaveFile( char *fn, int32_t *buff, int32_t size );
+int LoadFile( char *fn, int32_t *buff, int32_t size );
+int32_t GetConfig( char* fn, char* cParam );
+int32_t LogFileWrite( char* fn, char* cParam );
 
-long LoadBitmap( char *fname , int bmpindex, int flag );
+int32_t LoadBitmap( char *fname , int bmpindex, int flag );
 void ReleaseBitmap( int bmpindex );
 void BltRect(int bmpindex, int srcX, int srcY, int dstX, int dstY, int width, int height);
 void Blt( int bmpindex, int dstX, int dstY );
 void ClearSecondary( void );
-void BltNumericImage( long value, long length, long x, long y, long plane, long num_stpos_x, long num_stpos_y, long num_width, long num_height );
-void BltNumericImage2( long value, long length, long x, long y, long plane, long num_stpos_x, long num_stpos_y, long num_width, long num_height );
-long get2keta( long val, long st );
+void BltNumericImage( int32_t value, int32_t length, int32_t x, int32_t y, int32_t plane, int32_t num_stpos_x, int32_t num_stpos_y, int32_t num_width, int32_t num_height );
+void BltNumericImage2( int32_t value, int32_t length, int32_t x, int32_t y, int32_t plane, int32_t num_stpos_x, int32_t num_stpos_y, int32_t num_width, int32_t num_height );
+int32_t get2keta( int32_t val, int32_t st );
 void SetGscreenPalette( SDL_Surface *surface );
 void SetPalette(int getbmpindex, int setbmpindex);
 void BltRectRotZoom( int bmpindex, int dstX, int dstY, int srcX, int srcY, int width, int height, double angle, double zoom, int smooth);
 void CreateSurface( int bmpindex, int size_x, int size_y  );
 void SwapToSecondary( int bmpindex );
 void SaveBmp( int bmpindex, char *fn );
-void drawGRPline(f32 x1, f32 y1, f32 x2, f32 y2, Uint32 color);
-void pointSDLsurface( f32 px, f32 py, Uint32 color);
+void drawGRPline(int32_t x1, int32_t y1, int32_t x2, int32_t y2, Uint32 color);
+void pointSDLsurface( int32_t px, int32_t py, Uint32 color);
 void putSDLpixel(SDL_Surface *surface, int x, int y, Uint32 pixel);
 
 int IsPushKey( int keycode );
@@ -62,9 +62,9 @@ void soundInitBuffer(void);
 void soundRelease(void);
 void soundLoadBuffer(Sint32 num, Uint8 *fname, int loop);
 
-long funcSin( long rdo );
-long funcCos( long rdo );
-long funcTan2( long posX, long posY );
+int32_t funcSin( int32_t rdo );
+int32_t funcCos( int32_t rdo );
+int32_t funcTan2( int32_t posX, int32_t posY );
 
 // キー取得用
 static int key_eventPress[GP2X_BUTTON_MAX];
@@ -85,8 +85,8 @@ Uint8 *keys;
 static SDL_Surface* bitmap[BMPBUFF_MAX];
 static SDL_Surface* g_surface_bakup;
 // 定時処理用
-static long prvTickCount;
-static long nowTick;
+static int32_t prvTickCount;
+static int32_t nowTick;
 //static int frame;
 
 #ifdef GP2X
@@ -98,7 +98,7 @@ static const int INTERVAL_BASE = 16;
 SDL_Event event;
 
 // 音楽再生
-static long sound_vol;
+static int32_t sound_vol;
 
 void FunctionInit( void )
 {
@@ -312,7 +312,7 @@ int SaveGameFlag2( char *fn )
 	
 	return ( rc );
 }
-int SaveFile( char *fn, long *buff, long size )
+int SaveFile( char *fn, int32_t *buff, int32_t size )
 {
 	int rc = 0;
 	FILE *fp;	
@@ -343,7 +343,7 @@ int SaveFile( char *fn, long *buff, long size )
 	
 	return ( rc );
 }
-int LoadFile( char *fn, long *buff, long size )
+int LoadFile( char *fn, int32_t *buff, int32_t size )
 {
 	int rc = 0;
 	FILE *fp;	
@@ -375,14 +375,14 @@ int LoadFile( char *fn, long *buff, long size )
 	return ( rc );
 }
 
-long GetConfig( char* fn, char* cParam )
+int32_t GetConfig( char* fn, char* cParam )
 {
 	FILE *fp;	
 	char *sp;
 	char s[256];
 	char s2[256];
 	memset( s, '\0', sizeof( s ) );
-	long rc;
+	int32_t rc;
 	
 	rc = 0;
 
@@ -433,10 +433,10 @@ long GetConfig( char* fn, char* cParam )
 	return( rc );
 }
 
-long LogFileWrite( char* fn, char* cParam )
+int32_t LogFileWrite( char* fn, char* cParam )
 {
 	FILE *fp;	
-	long rc;
+	int32_t rc;
 	
 	rc = 0;
 
@@ -465,9 +465,9 @@ long LogFileWrite( char* fn, char* cParam )
 }
 
 
-long LoadBitmap( char *fname , int bmpindex, int flag )
+int32_t LoadBitmap( char *fname , int bmpindex, int flag )
 {
-	long rc;
+	int32_t rc;
 	SDL_Surface* tmp;
 	char filename[1024];
 
@@ -692,9 +692,9 @@ void ClearSecondary( void )
 	SDL_FillRect(g_screen, &rect, SDL_MapRGBA(g_screen->format,0,0,0,255));
 }
 
-void drawGRPline(f32 x1, f32 y1, f32 x2, f32 y2, Uint32 color)
+void drawGRPline(int32_t x1, int32_t y1, int32_t x2, int32_t y2, Uint32 color)
 {
-	f32 x, y, dx, dy, s, step;
+	int32_t x, y, dx, dy, s, step;
 	int i, j;
 
 	dx = abs((x2 >> 16) - (x1 >> 16)) * WP;
@@ -784,7 +784,7 @@ void drawGRPline(f32 x1, f32 y1, f32 x2, f32 y2, Uint32 color)
 		}
 	}
 }
-void pointSDLsurface( f32 px, f32 py, Uint32 color)
+void pointSDLsurface( int32_t px, int32_t py, Uint32 color)
 {
 	if ( px < 0 || px >= ( DISPLY_WIDTH_WP ) || py < 0 || py >= ( DISPLY_HEIGHT_WP ) )
 	{
@@ -1291,11 +1291,11 @@ int Set_Volume( int vol )
 /*-[戻り値]---------------------------------------------------------*/
 /*	無し															*/
 
-void BltNumericImage( long value, long length, long x, long y, long plane, long num_stpos_x, long num_stpos_y, long num_width, long num_height )
+void BltNumericImage( int32_t value, int32_t length, int32_t x, int32_t y, int32_t plane, int32_t num_stpos_x, int32_t num_stpos_y, int32_t num_width, int32_t num_height )
 {
-	long blt_num;	// １桁の数値を格納する
-	long i;			// 桁数分のforループで使用
-	long dv;		// 割り算で使用する値
+	int32_t blt_num;	// １桁の数値を格納する
+	int32_t i;			// 桁数分のforループで使用
+	int32_t dv;		// 割り算で使用する値
 
 	// value が負の値の場合、正の値に置き換える
 	if ( value < 0 )
@@ -1344,12 +1344,12 @@ void BltNumericImage( long value, long length, long x, long y, long plane, long 
 /*-[戻り値]---------------------------------------------------------*/
 /*	無し															*/
 /********************************************************************/
-void BltNumericImage2( long value, long length, long x, long y, long plane, long num_stpos_x, long num_stpos_y, long num_width, long num_height )
+void BltNumericImage2( int32_t value, int32_t length, int32_t x, int32_t y, int32_t plane, int32_t num_stpos_x, int32_t num_stpos_y, int32_t num_width, int32_t num_height )
 {
-	long blt_num;	// １桁の数値を格納する
-	long i;			// 桁数分のforループで使用
-	long dv;		// 割り算で使用する値
-	long x_hosei;	//右詰め補正値
+	int32_t blt_num;	// １桁の数値を格納する
+	int32_t i;			// 桁数分のforループで使用
+	int32_t dv;		// 割り算で使用する値
+	int32_t x_hosei;	//右詰め補正値
 	
 	// value が負の値の場合、正の値に置き換える
 	if ( value < 0 )
@@ -1408,10 +1408,10 @@ void BltNumericImage2( long value, long length, long x, long y, long plane, long
 	return;
 }
 
-long funcSin( long rdo )
+int32_t funcSin( int32_t rdo )
 {
-	long ang = 0;
-	long rc = 0;
+	int32_t ang = 0;
+	int32_t rc = 0;
 	
 	if ( ( rdo >= 0 ) && ( rdo < 180 ) )
 	{
@@ -1428,10 +1428,10 @@ long funcSin( long rdo )
 	
 	return( rc );
 }
-long funcCos( long rdo )
+int32_t funcCos( int32_t rdo )
 {
-	long ang = 0;
-	long rc = 0;
+	int32_t ang = 0;
+	int32_t rc = 0;
 	
 	if ( ( rdo >= 0 ) && ( rdo < 180 ) )
 	{
@@ -1450,18 +1450,18 @@ long funcCos( long rdo )
 }
 
 
-long funcTan2( long posX, long posY )
+int32_t funcTan2( int32_t posX, int32_t posY )
 {
-	long rc = 0;
+	int32_t rc = 0;
 	
 	rc = MOTatan( ( posX ) * 256, ( posY * -1 ) * 256 );
 	
 	return( rc );
 }
 
-long get2keta( long val, long st )
+int32_t get2keta( int32_t val, int32_t st )
 {
-	long rc = 0;
+	int32_t rc = 0;
 	
 	val = val / st;
 	rc = val % 100;
